+++ packages/0g/src/filters.ts
@@ -1,13 +1,13 @@
-import { ComponentType } from './Component.js';
+import { ComponentHandle } from './Component2.js';
 
-export type Has<Comp extends ComponentType<any>> = {
+export type Has<Comp extends ComponentHandle> = {
   Component: Comp;
   kind: 'has';
   __isFilter: true;
   toString(): string;
 };
 
-export const has = <Comp extends ComponentType<any>>(
+export const has = <Comp extends ComponentHandle>(
   Component: Comp,
 ): Has<Comp> => ({
   Component,
@@ -18,14 +18,14 @@ export const has = <Comp extends ComponentType<any>>(
   },
 });
 
-export type Not<Comp extends ComponentType<any>> = {
+export type Not<Comp extends ComponentHandle> = {
   Component: Comp;
   kind: 'not';
   __isFilter: true;
   toString(): string;
 };
 
-export const not = <Comp extends ComponentType<any>>(
+export const not = <Comp extends ComponentHandle>(
   Component: Comp,
 ): Not<Comp> => ({
   Component,
@@ -36,14 +36,14 @@ export const not = <Comp extends ComponentType<any>>(
   },
 });
 
-export type Changed<Comp extends ComponentType<any>> = {
+export type Changed<Comp extends ComponentHandle> = {
   Component: Comp;
   kind: 'changed';
   __isFilter: true;
   toString(): string;
 };
 
-export const changed = <Comp extends ComponentType<any>>(
+export const changed = <Comp extends ComponentHandle>(
   Component: Comp,
 ): Changed<Comp> => ({
   Component,
@@ -54,14 +54,14 @@ export const changed = <Comp extends ComponentType<any>>(
   },
 });
 
-export type Any<Comps extends ComponentType<any>[]> = {
+export type Any<Comps extends ComponentHandle[]> = {
   Components: Comps;
   kind: 'any';
   __isFilter: true;
   toString(): string;
 };
 
-export const any = <Comps extends ComponentType<any>[]>(
+export const any = <Comps extends ComponentHandle[]>(
   ...Components: Comps
 ): Any<Comps> => ({
   Components,
@@ -72,7 +72,7 @@ export const any = <Comps extends ComponentType<any>[]>(
   },
 });
 
-export type Filter<Comp extends ComponentType<any>> =
+export type Filter<Comp extends ComponentHandle> =
   | Not<Comp>
   | Has<Comp>
   | Changed<Comp>
@@ -81,16 +81,15 @@ export type Filter<Comp extends ComponentType<any>> =
 export const isFilter = (thing: any): thing is Filter<any> =>
   thing.__isFilter === true;
 
-export const isNotFilter = (fil: Filter<any>): fil is Not<ComponentType<any>> =>
+export const isNotFilter = (fil: Filter<any>): fil is Not<ComponentHandle> =>
   fil.kind === 'not';
 
-export const isHasFilter = (fil: Filter<any>): fil is Has<ComponentType<any>> =>
+export const isHasFilter = (fil: Filter<any>): fil is Has<ComponentHandle> =>
   fil.kind === 'has';
 
 export const isChangedFilter = (
   fil: Filter<any>,
-): fil is Changed<ComponentType<any>> => fil.kind === 'changed';
+): fil is Changed<ComponentHandle> => fil.kind === 'changed';
 
-export const isAnyFilter = (
-  fil: Filter<any>,
-): fil is Any<ComponentType<any>[]> => fil.kind === 'any';
+export const isAnyFilter = (fil: Filter<any>): fil is Any<ComponentHandle[]> =>
+  fil.kind === 'any';