+++ packages/react/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @0g/react
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [911246b]
+  - 0g@0.2.1
+
+## 2.0.0
+
+### Minor Changes
+
+- 72203f1: Rewrite components as plain objects
+
+### Patch Changes
+
+- Updated dependencies [72203f1]
+  - 0g@0.2.0
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [35733cd]
+  - 0g@0.1.2
+
 ## 1.0.1
 
 ### Patch Changes